@@ -4,12 +4,9 @@ version: 1.0.26+2
 author: Dylan Wu <dylanwuzh@gmail.com>
 homepage: https://github.com/wuzhendev/flutter-cupertino-date-picker
 
+environment:
+  sdk: '>=2.12.0 <3.0.0'
+
 dependencies:
   flutter:
     sdk: flutter
-
-flutter:
-
-environment:
-  sdk: ">=1.19.0 <3.0.0"
-  flutter: ">=0.4.1 <2.0.0"
