@@ -29,7 +29,7 @@ class _DatePickerInPageState extends State<DatePickerInPage> {
   @override
   Widget build(BuildContext context) {
     TextStyle hintTextStyle =
-        Theme.of(context).textTheme.subhead.apply(color: Color(0xFF999999));
+        Theme.of(context).textTheme.subtitle1.apply(color: Color(0xFF999999));
     return Scaffold(
       appBar: AppBar(title: Text("DatePicker In Page")),
       body: Container(
@@ -47,7 +47,7 @@ class _DatePickerInPageState extends State<DatePickerInPage> {
                     child: Text('min DateTime:', style: hintTextStyle),
                   ),
                   Text(MIN_DATETIME,
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -61,7 +61,7 @@ class _DatePickerInPageState extends State<DatePickerInPage> {
                       width: 115.0,
                       child: Text('max DateTime:', style: hintTextStyle)),
                   Text(MAX_DATETIME,
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -76,7 +76,7 @@ class _DatePickerInPageState extends State<DatePickerInPage> {
                     child: Text('init DateTime:', style: hintTextStyle),
                   ),
                   Text(INIT_DATETIME,
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -90,7 +90,8 @@ class _DatePickerInPageState extends State<DatePickerInPage> {
                     width: 115.0,
                     child: Text('Date Format:', style: hintTextStyle),
                   ),
-                  Text(DATE_FORMAT, style: Theme.of(context).textTheme.subhead),
+                  Text(DATE_FORMAT,
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -119,7 +120,7 @@ DateTimePickerTheme(
                       style: TextStyle(fontSize: 14.0),
                     ),
                     actions: <Widget>[
-                      new FlatButton(
+                      new TextButton(
                         child: new Text("OK"),
                         onPressed: () {
                           Navigator.of(context).pop();
@@ -173,14 +174,14 @@ DateTimePickerTheme(
               crossAxisAlignment: CrossAxisAlignment.center,
               children: <Widget>[
                 Text('Selected Date:',
-                    style: Theme.of(context).textTheme.subhead),
+                    style: Theme.of(context).textTheme.subtitle1),
                 Container(
                   padding: EdgeInsets.only(left: 12.0),
                   child: Text(
                     _dateTime != null
                         ? '${_dateTime.year}-${_dateTime.month.toString().padLeft(2, '0')}-${_dateTime.day.toString().padLeft(2, '0')}'
                         : '',
-                    style: Theme.of(context).textTheme.title,
+                    style: Theme.of(context).textTheme.headline6,
                   ),
                 ),
               ],
