@@ -18,7 +18,7 @@ const double DATETIME_PICKER_ITEM_HEIGHT = 36.0;
 
 /// Default value of DatePicker's item [TextStyle].
 const TextStyle DATETIME_PICKER_ITEM_TEXT_STYLE =
-    const TextStyle(color: Color(0xFF000046), fontSize: 16.0);
+    TextStyle(color: Color(0xFF000046), fontSize: 16.0);
 
 class DateTimePickerTheme with Diagnosticable {
   final cancelDefault = const Text('OK');
@@ -37,38 +37,38 @@ class DateTimePickerTheme with Diagnosticable {
   /// [itemHeight] The value of DatePicker's column height.
   /// [itemTextStyle] The value of DatePicker's column [TextStyle].
   const DateTimePickerTheme({
-    this.backgroundColor: DATETIME_PICKER_BACKGROUND_COLOR,
+    this.backgroundColor = DATETIME_PICKER_BACKGROUND_COLOR,
     this.cancelTextStyle,
     this.confirmTextStyle,
     this.cancel,
     this.confirm,
     this.title,
-    this.showTitle: DATETIME_PICKER_SHOW_TITLE_DEFAULT,
-    this.pickerHeight: DATETIME_PICKER_HEIGHT,
-    this.titleHeight: DATETIME_PICKER_TITLE_HEIGHT,
-    this.itemHeight: DATETIME_PICKER_ITEM_HEIGHT,
-    this.itemTextStyle: DATETIME_PICKER_ITEM_TEXT_STYLE,
+    this.showTitle = DATETIME_PICKER_SHOW_TITLE_DEFAULT,
+    this.pickerHeight = DATETIME_PICKER_HEIGHT,
+    this.titleHeight = DATETIME_PICKER_TITLE_HEIGHT,
+    this.itemHeight = DATETIME_PICKER_ITEM_HEIGHT,
+    this.itemTextStyle = DATETIME_PICKER_ITEM_TEXT_STYLE,
   });
 
-  static const DateTimePickerTheme Default = const DateTimePickerTheme();
+  static const DateTimePickerTheme Default = DateTimePickerTheme();
 
   /// DatePicker's background color.
   final Color backgroundColor;
 
   /// Default cancel widget's [TextStyle].
-  final TextStyle cancelTextStyle;
+  final TextStyle? cancelTextStyle;
 
   /// Default confirm widget's [TextStyle].
-  final TextStyle confirmTextStyle;
+  final TextStyle? confirmTextStyle;
 
   /// Custom cancel [Widget].
-  final Widget cancel;
+  final Widget? cancel;
 
   /// Custom confirm [Widget].
-  final Widget confirm;
+  final Widget? confirm;
 
   /// Custom title [Widget]. If specify a title widget, the cancel and confirm widgets will not display.
-  final Widget title;
+  final Widget? title;
 
   /// Whether display title widget or not. If set false, the default cancel and confirm widgets will not display, but the custom title widget will display if had specified one custom title widget.
   final bool showTitle;
