@@ -10,25 +10,25 @@ import '../i18n/date_picker_i18n.dart';
 /// @since 2019-05-16
 class DatePickerTitleWidget extends StatelessWidget {
   DatePickerTitleWidget({
-    Key key,
+    Key? key,
     this.pickerTheme,
     this.locale,
-    @required this.onCancel,
-    @required this.onConfirm,
+    required this.onCancel,
+    required this.onConfirm,
   }) : super(key: key);
 
-  final DateTimePickerTheme pickerTheme;
-  final DateTimePickerLocale locale;
+  final DateTimePickerTheme? pickerTheme;
+  final DateTimePickerLocale? locale;
   final DateVoidCallback onCancel, onConfirm;
 
   @override
   Widget build(BuildContext context) {
-    if (pickerTheme.title != null) {
-      return pickerTheme.title;
+    if (pickerTheme!.title != null) {
+      return pickerTheme!.title!;
     }
     return Container(
-      height: pickerTheme.titleHeight,
-      decoration: BoxDecoration(color: pickerTheme.backgroundColor),
+      height: pickerTheme!.titleHeight,
+      decoration: BoxDecoration(color: pickerTheme!.backgroundColor),
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: <Widget>[
@@ -43,14 +43,14 @@ class DatePickerTitleWidget extends StatelessWidget {
   Widget _renderCancelWidget(BuildContext context) {
     if (isCustomTitleWidget()) {
       // has custom title button widget
-      if (pickerTheme.cancel == null) {
+      if (pickerTheme!.cancel == null) {
         return Offstage();
       }
     }
 
-    Widget cancelWidget = pickerTheme.cancel;
+    Widget? cancelWidget = pickerTheme!.cancel;
     if (cancelWidget == null) {
-      TextStyle textStyle = pickerTheme.cancelTextStyle ??
+      TextStyle textStyle = pickerTheme!.cancelTextStyle ??
           TextStyle(
             color: Theme.of(context).unselectedWidgetColor,
             fontSize: 16.0,
@@ -62,9 +62,13 @@ class DatePickerTitleWidget extends StatelessWidget {
     }
 
     return Container(
-      height: pickerTheme.titleHeight,
-      child: FlatButton(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.zero),
+      height: pickerTheme!.titleHeight,
+      child: TextButton(
+        style: ButtonStyle(
+          shape: MaterialStateProperty.all(
+            RoundedRectangleBorder(borderRadius: BorderRadius.zero),
+          ),
+        ),
         child: cancelWidget,
         onPressed: () => this.onCancel(),
       ),
@@ -75,14 +79,14 @@ class DatePickerTitleWidget extends StatelessWidget {
   Widget _renderConfirmWidget(BuildContext context) {
     if (isCustomTitleWidget()) {
       // has custom title button widget
-      if (pickerTheme.confirm == null) {
+      if (pickerTheme!.confirm == null) {
         return Offstage();
       }
     }
 
-    Widget confirmWidget = pickerTheme.confirm;
+    Widget? confirmWidget = pickerTheme!.confirm;
     if (confirmWidget == null) {
-      TextStyle textStyle = pickerTheme.confirmTextStyle ??
+      TextStyle textStyle = pickerTheme!.confirmTextStyle ??
           TextStyle(
             color: Theme.of(context).primaryColor,
             fontSize: 16.0,
@@ -94,9 +98,13 @@ class DatePickerTitleWidget extends StatelessWidget {
     }
 
     return Container(
-      height: pickerTheme.titleHeight,
-      child: FlatButton(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.zero),
+      height: pickerTheme!.titleHeight,
+      child: TextButton(
+        style: ButtonStyle(
+          shape: MaterialStateProperty.all(
+            RoundedRectangleBorder(borderRadius: BorderRadius.zero),
+          ),
+        ),
         child: confirmWidget,
         onPressed: () => this.onConfirm(),
       ),
@@ -104,8 +112,8 @@ class DatePickerTitleWidget extends StatelessWidget {
   }
 
   bool isCustomTitleWidget() {
-    return pickerTheme.cancel != null ||
-        pickerTheme.confirm != null ||
-        pickerTheme.title != null;
+    return pickerTheme!.cancel != null ||
+        pickerTheme!.confirm != null ||
+        pickerTheme!.title != null;
   }
 }
