@@ -43,7 +43,8 @@ class _DateTimePickerBottomSheetState extends State<DateTimePickerBottomSheet> {
       pickerTheme: DateTimePickerTheme(
         showTitle: _showTitle,
       ),
-      pickerMode: DateTimePickerMode.datetime, // show TimePicker
+      pickerMode: DateTimePickerMode.datetime,
+      // show TimePicker
       onCancel: () {
         debugPrint('onCancel');
       },
@@ -62,31 +63,34 @@ class _DateTimePickerBottomSheetState extends State<DateTimePickerBottomSheet> {
 
   @override
   Widget build(BuildContext context) {
-    List<Widget> radios = List<Widget>();
-    _locales.forEach((locale) {
-      radios.add(Container(
+    // List<Widget> radios = List<Widget>();
+    List<Widget> radios = List.generate(
+      _locales.length,
+      (i) => Container(
         margin: EdgeInsets.only(right: 8.0),
         child: Row(
           mainAxisSize: MainAxisSize.min,
           children: <Widget>[
             Radio(
-                value: locale,
+                value: _locales[i],
                 groupValue: _locale,
                 onChanged: (value) {
                   setState(() {
                     _locale = value;
                   });
                 }),
-            Text(locale
-                .toString()
-                .substring(locale.toString().indexOf('.') + 1)),
+            Text(
+              _locales[i]
+                  .toString()
+                  .substring(_locales[i].toString().indexOf('.') + 1),
+            ),
           ],
         ),
-      ));
-    });
+      ),
+    );
 
     TextStyle hintTextStyle =
-        Theme.of(context).textTheme.subhead.apply(color: Color(0xFF999999));
+        Theme.of(context).textTheme.subtitle1.apply(color: Color(0xFF999999));
     return Scaffold(
       appBar: AppBar(title: Text('DateTimePicker Bottom Sheet')),
       body: Container(
@@ -103,7 +107,7 @@ class _DateTimePickerBottomSheetState extends State<DateTimePickerBottomSheet> {
                     child: Text('min DateTime:', style: hintTextStyle),
                   ),
                   Text(MIN_DATETIME,
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -118,7 +122,7 @@ class _DateTimePickerBottomSheetState extends State<DateTimePickerBottomSheet> {
                     child: Text('max DateTime:', style: hintTextStyle),
                   ),
                   Text(MAX_DATETIME,
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -133,7 +137,7 @@ class _DateTimePickerBottomSheetState extends State<DateTimePickerBottomSheet> {
                     child: Text('init DateTime:', style: hintTextStyle),
                   ),
                   Text(INIT_DATETIME,
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -188,12 +192,12 @@ class _DateTimePickerBottomSheetState extends State<DateTimePickerBottomSheet> {
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: <Widget>[
                   Text('Selected DateTime:',
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                   Container(
                     padding: EdgeInsets.only(top: 4.0),
                     child: Text(
                       '${_dateTime.year}-${_dateTime.month.toString().padLeft(2, '0')}-${_dateTime.day.toString().padLeft(2, '0')} ${_dateTime.hour.toString().padLeft(2, '0')}:${_dateTime.minute.toString().padLeft(2, '0')}:${_dateTime.second.toString().padLeft(2, '0')}',
-                      style: Theme.of(context).textTheme.title,
+                      style: Theme.of(context).textTheme.headline6,
                     ),
                   ),
                 ],
