@@ -29,7 +29,7 @@ class _DateTimePickerInPageState extends State<DateTimePickerInPage> {
   @override
   Widget build(BuildContext context) {
     TextStyle hintTextStyle =
-        Theme.of(context).textTheme.subhead.apply(color: Color(0xFF999999));
+        Theme.of(context).textTheme.subtitle1.apply(color: Color(0xFF999999));
     return Scaffold(
       appBar: AppBar(title: Text("DateTimePicker In Page")),
       body: Container(
@@ -47,7 +47,7 @@ class _DateTimePickerInPageState extends State<DateTimePickerInPage> {
                     child: Text('min DateTime:', style: hintTextStyle),
                   ),
                   Text(MIN_DATETIME,
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -62,7 +62,7 @@ class _DateTimePickerInPageState extends State<DateTimePickerInPage> {
                     child: Text('max DateTime:', style: hintTextStyle),
                   ),
                   Text(MAX_DATETIME,
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -77,7 +77,7 @@ class _DateTimePickerInPageState extends State<DateTimePickerInPage> {
                     child: Text('init DateTime:', style: hintTextStyle),
                   ),
                   Text(INIT_DATETIME,
-                      style: Theme.of(context).textTheme.subhead),
+                      style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -91,7 +91,7 @@ class _DateTimePickerInPageState extends State<DateTimePickerInPage> {
                     width: 115.0,
                     child: Text('Date Format:', style: hintTextStyle),
                   ),
-                  Text(DATE_FORMAT, style: Theme.of(context).textTheme.subhead),
+                  Text(DATE_FORMAT, style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -113,7 +113,7 @@ class _DateTimePickerInPageState extends State<DateTimePickerInPage> {
                     margin: EdgeInsets.only(right: 8.0),
                     child: Text('custom title height:', style: hintTextStyle),
                   ),
-                  Text('40.0', style: Theme.of(context).textTheme.subhead),
+                  Text('40.0', style: Theme.of(context).textTheme.subtitle1),
                 ],
               ),
             ),
@@ -150,7 +150,7 @@ class _DateTimePickerInPageState extends State<DateTimePickerInPage> {
               crossAxisAlignment: CrossAxisAlignment.start,
               children: <Widget>[
                 Text('Selected DateTime:',
-                    style: Theme.of(context).textTheme.subhead),
+                    style: Theme.of(context).textTheme.subtitle1),
                 Container(
                   width: double.infinity,
                   padding: EdgeInsets.only(top: 4.0),
@@ -158,7 +158,7 @@ class _DateTimePickerInPageState extends State<DateTimePickerInPage> {
                     _dateTime != null
                         ? '${_dateTime.year}-${_dateTime.month.toString().padLeft(2, '0')}-${_dateTime.day.toString().padLeft(2, '0')} ${_dateTime.hour.toString().padLeft(2, '0')}:${_dateTime.minute.toString().padLeft(2, '0')}:${_dateTime.second.toString().padLeft(2, '0')}'
                         : '',
-                    style: Theme.of(context).textTheme.title,
+                    style: Theme.of(context).textTheme.headline6,
                   ),
                 ),
               ],
