@@ -46,12 +46,12 @@ class _StringsEs extends _StringsI18n {
   }
 
   @override
-  List<String> getWeeksShort() {
+  List<String>? getWeeksShort() {
     return null;
   }
 
   @override
-  List<String> getMonthsShort() {
+  List<String>? getMonthsShort() {
     // TODO: implement getMonthsShort
     return null;
   }
