@@ -33,7 +33,7 @@ class _StringsTr extends _StringsI18n {
   }
 
   @override
-   List<String> getMonthsShort() {
+  List<String> getMonthsShort() {
     return [
       "Ock.",
       "Şub.",
